Week 5: Programming Assignment 3
You are given a list L. Write a program to print first prime number encountered in the list L.(Treat numbers below and equal to 1 as non prime)

Input will be handled by us.

Input
[1,2,3,4,5,6,7,8,9]

output
2

Explanation
Since 2 is the first prime number is list L, therefor it is printed.

Signed-off-by: taalojarvi <sreedevan05@gmail.com>